requirements.txt
Flask==3.1.2
SQLAlchemy==2.0.43
Flask-SQLAlchemy==3.1.1
bcrypt==4.3.0
PyJWT==2.10.1
bandit